Fix settings JSON opener test compile
